requirements.txt
bcrypt==3.2.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Deprecated==1.2.12
dnspython==2.1.0
email-validator==1.1.2
environs==9.3.2
fastapi==0.63.0
fastapi-users==5.1.3
firebase==3.0.1
firebase-admin==4.5.3
gcloud==0.17.0
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
gunicorn==20.1.0
h11==0.12.0
httplib2==0.19.1
idna==2.10
jws==0.1.3
makefun==1.11.3
marshmallow==3.11.1
msgpack==1.0.2
oauth2client==3.0.0
packaging==20.9
passlib==1.7.4
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pydantic==1.8.1
PyJWT==2.0.1
pyparsing==2.4.7
Pyrebase==3.0.27
python-dotenv==0.17.0
python-jwt==2.0.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
wrapt==1.12.1
